chore: restructure & remove webvitals (redunant)
