+++ package.json
@@ -38,7 +38,6 @@
     "react-scripts": "4.0.3",
     "rxjs": "^6.6.7",
     "tslib": "^2.3.1",
-    "web-vitals": "^1.0.1",
     "web3": "^1.2.11"
   },
   "devDependencies": {
